requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
decorator==4.4.2
Django==4.1.6
django-celery-beat==2.5.0
django-cleanup==6.0.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
imageio==2.32.0
imageio-ffmpeg==0.4.9
isort==5.12.0
kombu==5.3.3
moviepy==1.0.3
numpy==1.26.1
Pillow==10.0.1
proglog==0.1.10
prompt-toolkit==3.0.39
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.9
